main.py refactor
