+++ cogs/__init__.py
@@ -1,1 +1,1 @@
-import dota,music,rep,roulette,voice
+import dota,music,rep,roulette,voice,reactionlight
