+++ main.py
@@ -6,7 +6,7 @@ import os
 import utils, time
 import asyncio
 import database
-from cogs import voice, rep, dota, roulette
+from cogs import voice, rep, dota, roulette, reactionlight
 import logging
 import requests
 from dotenv import load_dotenv
@@ -50,8 +50,6 @@ prefixes = ["s!"]
 bot = commands.Bot(command_prefix=prefixes, intents=intents)
 bot.http.user_agent = "Discord IOS"
 bot.remove_command('help')
-
-db = database.Database("reactionlight.db")
 
 def system_notification(data):
 	print(data)
@@ -119,25 +117,25 @@ async def ball(ctx, *arg):
 	
 @bot.command()
 async def weather(ctx, *args):
-    arg = " ".join(args)
-    if not arg:
-        embed=discord.Embed(description=":warning: Укажите город",colour=botcolour)
-        embed.set_footer(icon_url=logo, text=botname)
-        await ctx.message.reply(embed=embed)
-        return
-    try:
-        weather = get_weather(arg)
-        icon = ":cloud:" if weather.weather.status == "Clouds" else ":sunny:" if weather.weather.status == "Clear" else ":question:"
-        embed = discord.Embed(
-            description=f"{icon} В городе {weather.location.name} сейчас {weather.weather.detailed_status} ({weather.weather.status})",
-            colour=botcolour,
-        )
-        embed.set_footer(icon_url=logo, text=botname)
-        await ctx.message.reply(embed=embed)
-    except NotFoundError:
-    	embed=discord.Embed(description=":warning: Город \""+arg+"\" не найден!",colour=botcolour)
-    	embed.set_footer(icon_url=logo, text=botname)
-    	await ctx.message.reply(embed=embed)
+	arg = " ".join(args)
+	if not arg:
+		embed=discord.Embed(description=":warning: Укажите город",colour=botcolour)
+		embed.set_footer(icon_url=logo, text=botname)
+		await ctx.message.reply(embed=embed)
+		return
+	try:
+		weather = get_weather(arg)
+		icon = ":cloud:" if weather.weather.status == "Clouds" else ":sunny:" if weather.weather.status == "Clear" else ":question:"
+		embed = discord.Embed(
+			description=f"{icon} В городе {weather.location.name} сейчас {weather.weather.detailed_status} ({weather.weather.status})",
+			colour=botcolour,
+		)
+		embed.set_footer(icon_url=logo, text=botname)
+		await ctx.message.reply(embed=embed)
+	except NotFoundError:
+		embed=discord.Embed(description=":warning: Город \""+arg+"\" не найден!",colour=botcolour)
+		embed.set_footer(icon_url=logo, text=botname)
+		await ctx.message.reply(embed=embed)
 
 @bot.command()
 async def vip(ctx):
@@ -170,10 +168,11 @@ async def setup_cogs():
 	await bot.add_cog(voice.Voice(bot))
 	await bot.add_cog(rep.ReputationCog(bot))
 	await bot.add_cog(dota.DotaInfo(bot))
+	await bot.add_cog(reactionlight.ReactionLight(bot))
 #	await bot.add_cog(speech.Speech(bot))
 	#bot.add_cog(icon.ServerIcon(bot))
 #	mod.setup(bot)
-	await bot.add_cog(roulette.Game(bot))
+	#await bot.add_cog(roulette.Game(bot))
 	
 loop = asyncio.get_event_loop()
